Abstract class Benchmark
