+++ workloads/0%-10%/low_variability.csv
@@ -1,15 +1,15 @@
 filepath,num_joins_in_user_query,num_joins_in_benchmark_query,query_id
-imdb/benchmarks/ceb/10a/b0fa6a3f8fa1820a4f7be7aeb6b405909834053c.sql,5,6,36893
-imdb/benchmarks/ceb/10a/aa1883d2afa0b06a4d39a9ac7a59f64b70e056bd.sql,5,6,237356
-imdb/benchmarks/ceb/10a/7bcc6d0a7d4f71732a63dd7ed1fa90b920521629.sql,7,11,267373
-imdb/benchmarks/ceb/10a/91b99b194361afcf7e32ecc811548342a16e05aa.sql,5,6,225671
-imdb/benchmarks/ceb/10a/76606f72fdf1a6b058fa85061c5ebaaac1dd0b1f.sql,5,6,10367
-imdb/benchmarks/ceb/10a/6a8c485255dad8338337788b05933746cfba3a86.sql,5,6,235812
-imdb/benchmarks/ceb/10a/69392b806b2d245ebf7e4ae196b9947536c45af4.sql,5,6,269325
-imdb/benchmarks/ceb/10a/d4eec8b3940f97bddc1b30930d723b9b1245aee9.sql,5,6,360116
-imdb/benchmarks/ceb/10a/af3f46ab118a1c2d503aeeee2e8c5ae79687f8a1.sql,5,6,142997
-imdb/benchmarks/ceb/10a/4f19c56e1f0ccdeca925474a607fb0d9994532ff.sql,7,11,370706
-imdb/benchmarks/ceb/10a/f33c42be8454d0e1823e578572022699fb67185d.sql,5,6,208350
-imdb/benchmarks/ceb/10a/c512f5577fe1aa92710f4ec405e10515c496cd32.sql,5,6,84423
-imdb/benchmarks/ceb/10a/4b303f7a1aa4d1474bae2a43a7c6d7c251bb574a.sql,5,6,27229
-imdb/benchmarks/ceb/10a/ae55eb3c061489ee5263ca7886a443b40defdfe8.sql,5,6,228856
+imdb/benchmarks/ceb/10a/ffa89001847857aea332203f58817579edf17efa.sql,5,6,36893
+imdb/benchmarks/ceb/10a/ff7e187ff73ce227ca49c7b4791037f242cdbde6.sql,5,6,237356
+imdb/benchmarks/ceb/10a/ff5392412c047db440be1c25a9368dca0af322a6.sql,7,11,267373
+imdb/benchmarks/ceb/10a/ff28e87026770d06e47c8133bde907e8f62af5f2.sql,5,6,225671
+imdb/benchmarks/ceb/10a/feb4a84432cbf3c03659dec9f994e45c20117dce.sql,5,6,10367
+imdb/benchmarks/ceb/10a/feab1842e30c452abd391c3326d9e4ff34c6dc11.sql,5,6,235812
+imdb/benchmarks/ceb/10a/fe908e2b6537a8d6acfd277573458dce0e168e28.sql,5,6,269325
+imdb/benchmarks/ceb/10a/fe466f77ec0cd3dca2c428d45227925db297555a.sql,5,6,360116
+imdb/benchmarks/ceb/10a/fe2eb3a8d76a12f5a5487e0a328bd294353322ca.sql,5,6,142997
+imdb/benchmarks/ceb/10a/fde055219af175b5eb19a7dc52449920e4abac50.sql,7,11,370706
+imdb/benchmarks/ceb/10a/fd4b8a0b054d271c165b33f400cbe960df4eaeaf.sql,5,6,208350
+imdb/benchmarks/ceb/10a/fd4595b17e8893f1919f57c23633b6707a56ee56.sql,5,6,84423
+imdb/benchmarks/ceb/10a/fd26fe6762a3823baca81ef3c51ead7cf3c5e955.sql,5,6,27229
+imdb/benchmarks/ceb/10a/fd0058725babf4fba5445b0226a9030da31f9aa8.sql,5,6,228856
