+++ workloads/0%-10%/mid_variability.csv
@@ -1,10 +1,10 @@
 filepath,num_joins_in_user_query,num_joins_in_benchmark_query,query_id
-imdb/benchmarks/ceb/8a/8a458.sql,7,11,2490412
-imdb/benchmarks/ceb/10a/b0fa6a3f8fa1820a4f7be7aeb6b405909834053c.sql,2,6,2408702
-imdb/benchmarks/ceb/10a/aa1883d2afa0b06a4d39a9ac7a59f64b70e056bd.sql,2,6,3381845
-imdb/benchmarks/ceb/10a/7bcc6d0a7d4f71732a63dd7ed1fa90b920521629.sql,2,6,3051348
-imdb/benchmarks/ceb/10a/91b99b194361afcf7e32ecc811548342a16e05aa.sql,2,6,2644680
-imdb/benchmarks/ceb/10a/76606f72fdf1a6b058fa85061c5ebaaac1dd0b1f.sql,2,6,1846138
+imdb/benchmarks/ceb/8a/8a99.sql,7,11,2490412
+imdb/benchmarks/ceb/10a/ffa89001847857aea332203f58817579edf17efa.sql,2,6,2408702
+imdb/benchmarks/ceb/10a/ff7e187ff73ce227ca49c7b4791037f242cdbde6.sql,2,6,3381845
+imdb/benchmarks/ceb/10a/ff5392412c047db440be1c25a9368dca0af322a6.sql,2,6,3051348
+imdb/benchmarks/ceb/10a/ff28e87026770d06e47c8133bde907e8f62af5f2.sql,2,6,2644680
+imdb/benchmarks/ceb/10a/feb4a84432cbf3c03659dec9f994e45c20117dce.sql,2,6,1846138
 imdb/benchmarks/job/11d.sql,3,7,2872114
-imdb/benchmarks/job/9a.sql,3,7,2787784
-imdb/benchmarks/ceb/10a/6a8c485255dad8338337788b05933746cfba3a86.sql,2,6,3437825
+imdb/benchmarks/job/16d.sql,3,7,2787784
+imdb/benchmarks/ceb/10a/feab1842e30c452abd391c3326d9e4ff34c6dc11.sql,2,6,3437825
