+++ workloads/10%-20%/low_variability.csv
@@ -1,6 +1,6 @@
 filepath,num_joins_in_user_query,num_joins_in_benchmark_query,query_id
-imdb/benchmarks/ceb/10a/b0fa6a3f8fa1820a4f7be7aeb6b405909834053c.sql,2,6,69718
-imdb/benchmarks/ceb/10a/aa1883d2afa0b06a4d39a9ac7a59f64b70e056bd.sql,3,11,133590
-imdb/benchmarks/ceb/10a/7bcc6d0a7d4f71732a63dd7ed1fa90b920521629.sql,3,11,134226
-imdb/benchmarks/ceb/10a/7bcc6d0a7d4f71732a63dd7ed1fa90b920521629.sql,3,11,173011
-imdb/benchmarks/ceb/10a/91b99b194361afcf7e32ecc811548342a16e05aa.sql,3,11,173782
+imdb/benchmarks/ceb/10a/ffa89001847857aea332203f58817579edf17efa.sql,2,6,69718
+imdb/benchmarks/ceb/10a/ff7e187ff73ce227ca49c7b4791037f242cdbde6.sql,3,11,133590
+imdb/benchmarks/ceb/10a/ff5392412c047db440be1c25a9368dca0af322a6.sql,3,11,134226
+imdb/benchmarks/ceb/10a/ff5392412c047db440be1c25a9368dca0af322a6.sql,3,11,173011
+imdb/benchmarks/ceb/10a/ff28e87026770d06e47c8133bde907e8f62af5f2.sql,3,11,173782
