+++ workloads/10%-20%/mid_variability.csv
@@ -1,17 +1,17 @@
 filepath,num_joins_in_user_query,num_joins_in_benchmark_query,query_id
-imdb/benchmarks/ceb/8a/8a458.sql,7,11,738428
-imdb/benchmarks/ceb/8a/8a458.sql,7,11,597850
-imdb/benchmarks/ceb/8a/8a377.sql,7,11,691792
-imdb/benchmarks/ceb/8a/8a283.sql,7,11,619643
-imdb/benchmarks/ceb/8a/8a478.sql,7,11,595728
-imdb/benchmarks/ceb/8a/8a283.sql,7,11,742064
-imdb/benchmarks/ceb/8a/8a391.sql,7,11,731099
-imdb/benchmarks/ceb/8a/8a88.sql,7,11,743630
-imdb/benchmarks/ceb/8a/8a283.sql,7,11,682868
-imdb/benchmarks/ceb/8a/8a37.sql,7,11,731664
-imdb/benchmarks/ceb/10a/b0fa6a3f8fa1820a4f7be7aeb6b405909834053c.sql,1,6,739377
+imdb/benchmarks/ceb/8a/8a99.sql,7,11,738428
+imdb/benchmarks/ceb/8a/8a99.sql,7,11,597850
+imdb/benchmarks/ceb/8a/8a98.sql,7,11,691792
+imdb/benchmarks/ceb/8a/8a97.sql,7,11,619643
+imdb/benchmarks/ceb/8a/8a96.sql,7,11,595728
+imdb/benchmarks/ceb/8a/8a97.sql,7,11,742064
+imdb/benchmarks/ceb/8a/8a95.sql,7,11,731099
+imdb/benchmarks/ceb/8a/8a94.sql,7,11,743630
+imdb/benchmarks/ceb/8a/8a97.sql,7,11,682868
+imdb/benchmarks/ceb/8a/8a93.sql,7,11,731664
+imdb/benchmarks/ceb/10a/ffa89001847857aea332203f58817579edf17efa.sql,1,6,739377
 imdb/benchmarks/job/11d.sql,2,7,635054
 imdb/benchmarks/job/11c.sql,2,7,701382
-imdb/benchmarks/job/11a.sql,2,7,686818
-imdb/benchmarks/ceb/1a/1a1623.sql,3,8,734585
-imdb/benchmarks/ceb/1a/1a1904.sql,3,8,624331
+imdb/benchmarks/job/11b.sql,2,7,686818
+imdb/benchmarks/ceb/1a/1a999.sql,3,8,734585
+imdb/benchmarks/ceb/1a/1a998.sql,3,8,624331
