+++ workloads/20%-30%/low_variability.csv
@@ -1,13 +1,13 @@
 filepath,num_joins_in_user_query,num_joins_in_benchmark_query,query_id
-imdb/benchmarks/ceb/8a/8a458.sql,3,11,3809964
-imdb/benchmarks/ceb/8a/8a377.sql,1,6,4792388
-imdb/benchmarks/ceb/8a/8a283.sql,1,6,4112837
-imdb/benchmarks/ceb/8a/8a478.sql,1,6,5750661
-imdb/benchmarks/ceb/8a/8a478.sql,1,6,5803323
-imdb/benchmarks/ceb/8a/8a391.sql,1,6,3769467
-imdb/benchmarks/ceb/8a/8a391.sql,1,6,3799736
-imdb/benchmarks/ceb/8a/8a88.sql,1,6,6791767
-imdb/benchmarks/ceb/8a/8a37.sql,1,6,4448196
-imdb/benchmarks/ceb/8a/8a37.sql,1,6,4521928
-imdb/benchmarks/ceb/8a/8a299.sql,1,6,6661093
-imdb/benchmarks/ceb/8a/8a242.sql,1,6,5875307
+imdb/benchmarks/ceb/8a/8a99.sql,3,11,3809964
+imdb/benchmarks/ceb/8a/8a98.sql,1,6,4792388
+imdb/benchmarks/ceb/8a/8a97.sql,1,6,4112837
+imdb/benchmarks/ceb/8a/8a96.sql,1,6,5750661
+imdb/benchmarks/ceb/8a/8a96.sql,1,6,5803323
+imdb/benchmarks/ceb/8a/8a95.sql,1,6,3769467
+imdb/benchmarks/ceb/8a/8a95.sql,1,6,3799736
+imdb/benchmarks/ceb/8a/8a94.sql,1,6,6791767
+imdb/benchmarks/ceb/8a/8a93.sql,1,6,4448196
+imdb/benchmarks/ceb/8a/8a93.sql,1,6,4521928
+imdb/benchmarks/ceb/8a/8a92.sql,1,6,6661093
+imdb/benchmarks/ceb/8a/8a91.sql,1,6,5875307
