+++ workloads/20%-30%/mid_variability.csv
@@ -1,26 +1,26 @@
 filepath,num_joins_in_user_query,num_joins_in_benchmark_query,query_id
-imdb/benchmarks/ceb/1a/1a1623.sql,2,8,1672184
-imdb/benchmarks/ceb/8a/8a458.sql,4,11,1672246
-imdb/benchmarks/ceb/8a/8a377.sql,4,11,2445543
-imdb/benchmarks/ceb/8a/8a377.sql,4,11,2927592
-imdb/benchmarks/ceb/8a/8a283.sql,4,11,3585476
-imdb/benchmarks/ceb/8a/8a377.sql,4,11,3712345
-imdb/benchmarks/ceb/8a/8a478.sql,4,11,1760827
-imdb/benchmarks/ceb/8a/8a478.sql,4,11,2913231
-imdb/benchmarks/ceb/8a/8a391.sql,4,11,1939260
-imdb/benchmarks/ceb/8a/8a88.sql,4,11,2740899
-imdb/benchmarks/ceb/8a/8a37.sql,4,11,3659978
-imdb/benchmarks/ceb/8a/8a299.sql,4,11,3287105
-imdb/benchmarks/ceb/8a/8a242.sql,4,11,2497754
-imdb/benchmarks/ceb/11b/703ebc6f215f2cfc1a93b38a7df263820024c93f.sql,4,11,3665852
-imdb/benchmarks/ceb/11b/362e55b8f90070e7ad2e0436667b34b4e50f5ec6.sql,4,11,3441679
-imdb/benchmarks/job/27b.sql,4,11,1668978
-imdb/benchmarks/job/27a.sql,4,11,2198122
-imdb/benchmarks/job/27c.sql,4,11,3586055
+imdb/benchmarks/ceb/1a/1a999.sql,2,8,1672184
+imdb/benchmarks/ceb/8a/8a99.sql,4,11,1672246
+imdb/benchmarks/ceb/8a/8a98.sql,4,11,2445543
+imdb/benchmarks/ceb/8a/8a98.sql,4,11,2927592
+imdb/benchmarks/ceb/8a/8a97.sql,4,11,3585476
+imdb/benchmarks/ceb/8a/8a98.sql,4,11,3712345
+imdb/benchmarks/ceb/8a/8a96.sql,4,11,1760827
+imdb/benchmarks/ceb/8a/8a96.sql,4,11,2913231
+imdb/benchmarks/ceb/8a/8a95.sql,4,11,1939260
+imdb/benchmarks/ceb/8a/8a94.sql,4,11,2740899
+imdb/benchmarks/ceb/8a/8a93.sql,4,11,3659978
+imdb/benchmarks/ceb/8a/8a92.sql,4,11,3287105
+imdb/benchmarks/ceb/8a/8a91.sql,4,11,2497754
+imdb/benchmarks/ceb/11b/fff7b3029c1269dafc8a42b3a1cf7cc0e994fd15.sql,4,11,3665852
+imdb/benchmarks/ceb/11b/fce5c5de24dbf6034bb0b4698816bc2b19670dce.sql,4,11,3441679
+imdb/benchmarks/job/26c.sql,4,11,1668978
+imdb/benchmarks/job/26b.sql,4,11,2198122
+imdb/benchmarks/job/26a.sql,4,11,3586055
 imdb/benchmarks/job/24b.sql,4,11,2927079
-imdb/benchmarks/job/26c.sql,4,11,2348543
-imdb/benchmarks/job/26c.sql,4,11,2218069
-imdb/benchmarks/job/26c.sql,4,11,1885091
-imdb/benchmarks/ceb/10a/b0fa6a3f8fa1820a4f7be7aeb6b405909834053c.sql,1,6,2551910
-imdb/benchmarks/ceb/10a/aa1883d2afa0b06a4d39a9ac7a59f64b70e056bd.sql,1,6,3091057
-imdb/benchmarks/job/30a.sql,4,11,3734199
+imdb/benchmarks/job/27c.sql,4,11,2348543
+imdb/benchmarks/job/27c.sql,4,11,2218069
+imdb/benchmarks/job/27c.sql,4,11,1885091
+imdb/benchmarks/ceb/10a/ffa89001847857aea332203f58817579edf17efa.sql,1,6,2551910
+imdb/benchmarks/ceb/10a/ff7e187ff73ce227ca49c7b4791037f242cdbde6.sql,1,6,3091057
+imdb/benchmarks/job/30c.sql,4,11,3734199
