+++ workloads/20%-30%/stats.csv
@@ -1,4 +1,4 @@
 workload_type,user_id,instance_id,number_of_queries,query_repetition_rate,n_distinct_num_joins,n_distinct_readsets,n_occurrences_step_2,n_occurrences_step_5,n_occurrences_step_3_to_6,n_occurrences_step_3_to_7,n_occurrences_step_4_to_6,n_occurrences_step_4_to_7
 low_variability,22,79,12,0.250,2,1,8,1,0,0,0,0
 mid_variability,203,33,25,0.200,3,7,12,7,1,0,0,0
-high_variability,0,132,1000,0.209,12,611,76,29,34,3,642,7
+high_variability,0,132,1000,0.209,12,611,77,28,33,2,643,8
