+++ workloads/30%-40%/high_variability.csv
@@ -1,63 +1,63 @@
 filepath,num_joins_in_user_query,num_joins_in_benchmark_query,query_id
 imdb/benchmarks/job/11d.sql,4,7,2140439
-imdb/benchmarks/job/9a.sql,4,7,4728914
-imdb/benchmarks/ceb/3a/3a710.sql,8,9,4970543
-imdb/benchmarks/job/16b.sql,4,7,3654500
-imdb/benchmarks/ceb/8a/8a458.sql,11,11,5021168
-imdb/benchmarks/ceb/2a/2a76.sql,9,10,3055404
-imdb/benchmarks/ceb/11b/703ebc6f215f2cfc1a93b38a7df263820024c93f.sql,11,11,4375551
-imdb/benchmarks/ceb/2b/2b81.sql,9,10,5355181
+imdb/benchmarks/job/16d.sql,4,7,4728914
+imdb/benchmarks/ceb/3a/3a999.sql,8,9,4970543
+imdb/benchmarks/job/9d.sql,4,7,3654500
+imdb/benchmarks/ceb/8a/8a99.sql,11,11,5021168
+imdb/benchmarks/ceb/2a/2a99.sql,9,10,3055404
+imdb/benchmarks/ceb/11b/fff7b3029c1269dafc8a42b3a1cf7cc0e994fd15.sql,11,11,4375551
+imdb/benchmarks/ceb/2b/2b99.sql,9,10,5355181
 imdb/benchmarks/job/22d.sql,10,10,4233261
-imdb/benchmarks/ceb/5a/5a122.sql,8,9,5296838
+imdb/benchmarks/ceb/5a/5a999.sql,8,9,5296838
 imdb/benchmarks/job/23c.sql,10,10,5346575
-imdb/benchmarks/ceb/11a/a3d5c36a47d47c7869d06b549170bb15f75004d8.sql,8,9,2487140
-imdb/benchmarks/job/14b.sql,4,7,4606904
-imdb/benchmarks/ceb/3a/3a1120.sql,8,9,1136990
-imdb/benchmarks/ceb/10a/b0fa6a3f8fa1820a4f7be7aeb6b405909834053c.sql,2,6,95584
-imdb/benchmarks/ceb/3a/3a1120.sql,8,9,1689746
-imdb/benchmarks/ceb/3b/529792bcfcec19c721f4ed7ce43096b415173571.sql,8,9,1970546
-imdb/benchmarks/ceb/1a/1a1623.sql,6,8,376926
-imdb/benchmarks/job/27b.sql,12,11,841919
-imdb/benchmarks/job/26c.sql,12,11,1318390
-imdb/benchmarks/job/30a.sql,12,11,1478641
-imdb/benchmarks/ceb/9a/8b7f2755bbf3f958ca6e4185ee969497c41fc3ac.sql,6,8,1521258
-imdb/benchmarks/ceb/9b/45e8cad34180083e6beb2fad21482ee924d03528.sql,6,8,1832031
+imdb/benchmarks/ceb/11a/fe9f99ecc087c8840691c9b100708bfa43569855.sql,8,9,2487140
+imdb/benchmarks/job/12c.sql,4,7,4606904
+imdb/benchmarks/ceb/3a/3a998.sql,8,9,1136990
+imdb/benchmarks/ceb/10a/ffa89001847857aea332203f58817579edf17efa.sql,2,6,95584
+imdb/benchmarks/ceb/3a/3a998.sql,8,9,1689746
+imdb/benchmarks/ceb/3b/ff942a1052c1a10be7e152823a06322c07a314a3.sql,8,9,1970546
+imdb/benchmarks/ceb/1a/1a999.sql,6,8,376926
+imdb/benchmarks/job/26c.sql,12,11,841919
+imdb/benchmarks/job/27c.sql,12,11,1318390
+imdb/benchmarks/job/30c.sql,12,11,1478641
+imdb/benchmarks/ceb/9a/fff673d500e25c73a6e054e9a4e60a31e633ba47.sql,6,8,1521258
+imdb/benchmarks/ceb/9b/feae2fd4e41ac4a2d94f9d77bf9d0106a7397871.sql,6,8,1832031
 imdb/benchmarks/job/24b.sql,11,11,1707758
-imdb/benchmarks/job/24a.sql,11,11,227055
-imdb/benchmarks/job/14c.sql,4,7,1594582
-imdb/benchmarks/job/31a.sql,10,10,1999451
-imdb/benchmarks/ceb/2a/2a214.sql,10,10,1417511
+imdb/benchmarks/job/26b.sql,11,11,227055
+imdb/benchmarks/job/12b.sql,4,7,1594582
+imdb/benchmarks/job/31c.sql,10,10,1999451
+imdb/benchmarks/job/23b.sql,10,10,1417511
 imdb/benchmarks/job/11c.sql,4,7,465091
-imdb/benchmarks/ceb/3a/3a1120.sql,8,9,407797
-imdb/benchmarks/ceb/10a/b0fa6a3f8fa1820a4f7be7aeb6b405909834053c.sql,2,6,169054
-imdb/benchmarks/ceb/3b/529792bcfcec19c721f4ed7ce43096b415173571.sql,8,9,1075666
-imdb/benchmarks/ceb/3a/3a1120.sql,8,9,1692619
-imdb/benchmarks/ceb/3b/529792bcfcec19c721f4ed7ce43096b415173571.sql,8,9,430140
-imdb/benchmarks/ceb/1a/1a1623.sql,6,8,2008252
-imdb/benchmarks/job/27b.sql,12,11,886643
-imdb/benchmarks/job/26c.sql,12,11,1818292
-imdb/benchmarks/job/30a.sql,12,11,559701
-imdb/benchmarks/ceb/9a/8b7f2755bbf3f958ca6e4185ee969497c41fc3ac.sql,6,8,1741931
-imdb/benchmarks/ceb/9b/45e8cad34180083e6beb2fad21482ee924d03528.sql,6,8,1785519
+imdb/benchmarks/ceb/3a/3a998.sql,8,9,407797
+imdb/benchmarks/ceb/10a/ffa89001847857aea332203f58817579edf17efa.sql,2,6,169054
+imdb/benchmarks/ceb/3b/ff942a1052c1a10be7e152823a06322c07a314a3.sql,8,9,1075666
+imdb/benchmarks/ceb/3a/3a998.sql,8,9,1692619
+imdb/benchmarks/ceb/3b/ff942a1052c1a10be7e152823a06322c07a314a3.sql,8,9,430140
+imdb/benchmarks/ceb/1a/1a999.sql,6,8,2008252
+imdb/benchmarks/job/26c.sql,12,11,886643
+imdb/benchmarks/job/27c.sql,12,11,1818292
+imdb/benchmarks/job/30c.sql,12,11,559701
+imdb/benchmarks/ceb/9a/fff673d500e25c73a6e054e9a4e60a31e633ba47.sql,6,8,1741931
+imdb/benchmarks/ceb/9b/feae2fd4e41ac4a2d94f9d77bf9d0106a7397871.sql,6,8,1785519
 imdb/benchmarks/job/24b.sql,11,11,1454088
-imdb/benchmarks/job/24a.sql,11,11,1456844
-imdb/benchmarks/job/14c.sql,4,7,806190
-imdb/benchmarks/job/31a.sql,10,10,672753
-imdb/benchmarks/ceb/2a/2a214.sql,10,10,1057990
-imdb/benchmarks/ceb/2c/2c25.sql,5,8,2039975
+imdb/benchmarks/job/26b.sql,11,11,1456844
+imdb/benchmarks/job/12b.sql,4,7,806190
+imdb/benchmarks/job/31c.sql,10,10,672753
+imdb/benchmarks/job/23b.sql,10,10,1057990
+imdb/benchmarks/ceb/2c/2c99.sql,5,8,2039975
 imdb/benchmarks/job/17f.sql,1,6,1389501
 imdb/benchmarks/job/17f.sql,1,6,1987033
 imdb/benchmarks/job/17f.sql,1,6,212775
-imdb/benchmarks/job/17a.sql,1,6,659213
-imdb/benchmarks/job/7c.sql,4,7,902995
-imdb/benchmarks/job/7a.sql,4,7,937579
-imdb/benchmarks/job/7b.sql,4,7,884094
-imdb/benchmarks/job/12c.sql,4,7,3368529
-imdb/benchmarks/job/12c.sql,4,7,2735586
-imdb/benchmarks/job/12c.sql,4,7,2432225
-imdb/benchmarks/job/12c.sql,4,7,3765527
-imdb/benchmarks/job/12a.sql,4,7,2765638
-imdb/benchmarks/job/12c.sql,4,7,3612206
-imdb/benchmarks/job/12b.sql,4,7,1978343
-imdb/benchmarks/job/12b.sql,4,7,2694455
-imdb/benchmarks/job/11a.sql,4,7,2661753
+imdb/benchmarks/job/17e.sql,1,6,659213
+imdb/benchmarks/job/14c.sql,4,7,902995
+imdb/benchmarks/job/14b.sql,4,7,937579
+imdb/benchmarks/job/14a.sql,4,7,884094
+imdb/benchmarks/job/7c.sql,4,7,3368529
+imdb/benchmarks/job/7c.sql,4,7,2735586
+imdb/benchmarks/job/7c.sql,4,7,2432225
+imdb/benchmarks/job/7c.sql,4,7,3765527
+imdb/benchmarks/job/7b.sql,4,7,2765638
+imdb/benchmarks/job/7c.sql,4,7,3612206
+imdb/benchmarks/job/7a.sql,4,7,1978343
+imdb/benchmarks/job/7a.sql,4,7,2694455
+imdb/benchmarks/job/11b.sql,4,7,2661753
