+++ workloads/30%-40%/mid_variability.csv
@@ -1,35 +1,35 @@
 filepath,num_joins_in_user_query,num_joins_in_benchmark_query,query_id
-imdb/benchmarks/ceb/3a/3a710.sql,2,9,226790
-imdb/benchmarks/ceb/10a/b0fa6a3f8fa1820a4f7be7aeb6b405909834053c.sql,1,6,503249
-imdb/benchmarks/ceb/3a/3a1120.sql,2,9,282935
-imdb/benchmarks/ceb/3a/3a1078.sql,2,9,354614
-imdb/benchmarks/ceb/8a/8a458.sql,3,11,532144
-imdb/benchmarks/ceb/8a/8a377.sql,3,11,170978
-imdb/benchmarks/ceb/3a/3a1078.sql,2,9,74764
-imdb/benchmarks/ceb/8a/8a458.sql,3,11,123844
-imdb/benchmarks/ceb/8a/8a458.sql,3,11,628021
-imdb/benchmarks/ceb/8a/8a458.sql,3,11,719005
-imdb/benchmarks/ceb/8a/8a283.sql,3,11,193311
-imdb/benchmarks/ceb/8a/8a478.sql,3,11,259100
-imdb/benchmarks/ceb/8a/8a391.sql,3,11,444369
-imdb/benchmarks/ceb/8a/8a88.sql,3,11,254549
-imdb/benchmarks/ceb/8a/8a37.sql,3,11,665811
-imdb/benchmarks/ceb/8a/8a299.sql,3,11,119379
-imdb/benchmarks/ceb/3a/3a1078.sql,2,9,482669
-imdb/benchmarks/ceb/3a/3a710.sql,2,9,534118
-imdb/benchmarks/ceb/3a/3a710.sql,2,9,399356
-imdb/benchmarks/ceb/8a/8a242.sql,3,11,536066
-imdb/benchmarks/ceb/8a/8a242.sql,3,11,251724
-imdb/benchmarks/ceb/8a/8a421.sql,3,11,553802
-imdb/benchmarks/ceb/8a/8a421.sql,3,11,304453
-imdb/benchmarks/ceb/8a/8a480.sql,3,11,576914
-imdb/benchmarks/ceb/8a/8a480.sql,3,11,670346
+imdb/benchmarks/ceb/3a/3a999.sql,2,9,226790
+imdb/benchmarks/ceb/10a/ffa89001847857aea332203f58817579edf17efa.sql,1,6,503249
+imdb/benchmarks/ceb/3a/3a998.sql,2,9,282935
+imdb/benchmarks/ceb/3a/3a997.sql,2,9,354614
+imdb/benchmarks/ceb/8a/8a99.sql,3,11,532144
+imdb/benchmarks/ceb/8a/8a98.sql,3,11,170978
+imdb/benchmarks/ceb/3a/3a997.sql,2,9,74764
+imdb/benchmarks/ceb/8a/8a99.sql,3,11,123844
+imdb/benchmarks/ceb/8a/8a99.sql,3,11,628021
+imdb/benchmarks/ceb/8a/8a99.sql,3,11,719005
+imdb/benchmarks/ceb/8a/8a97.sql,3,11,193311
+imdb/benchmarks/ceb/8a/8a96.sql,3,11,259100
+imdb/benchmarks/ceb/8a/8a95.sql,3,11,444369
+imdb/benchmarks/ceb/8a/8a94.sql,3,11,254549
+imdb/benchmarks/ceb/8a/8a93.sql,3,11,665811
+imdb/benchmarks/ceb/8a/8a92.sql,3,11,119379
+imdb/benchmarks/ceb/3a/3a997.sql,2,9,482669
+imdb/benchmarks/ceb/3a/3a999.sql,2,9,534118
+imdb/benchmarks/ceb/3a/3a999.sql,2,9,399356
+imdb/benchmarks/ceb/8a/8a91.sql,3,11,536066
+imdb/benchmarks/ceb/8a/8a91.sql,3,11,251724
+imdb/benchmarks/ceb/8a/8a90.sql,3,11,553802
+imdb/benchmarks/ceb/8a/8a90.sql,3,11,304453
+imdb/benchmarks/ceb/8a/8a9.sql,3,11,576914
+imdb/benchmarks/ceb/8a/8a9.sql,3,11,670346
 imdb/benchmarks/job/17f.sql,1,6,1555805
 imdb/benchmarks/job/17f.sql,1,6,1264896
-imdb/benchmarks/job/17a.sql,1,6,1307690
+imdb/benchmarks/job/17e.sql,1,6,1307690
 imdb/benchmarks/job/17d.sql,1,6,1468721
-imdb/benchmarks/job/17e.sql,1,6,1312070
+imdb/benchmarks/job/17c.sql,1,6,1312070
 imdb/benchmarks/job/17b.sql,1,6,1756351
-imdb/benchmarks/ceb/11b/703ebc6f215f2cfc1a93b38a7df263820024c93f.sql,3,11,1368287
-imdb/benchmarks/ceb/11b/362e55b8f90070e7ad2e0436667b34b4e50f5ec6.sql,3,11,1454015
-imdb/benchmarks/ceb/11b/9936ef125ccda1d03294df83fb45afb2e0b4e9a2.sql,3,11,1242451
+imdb/benchmarks/ceb/11b/fff7b3029c1269dafc8a42b3a1cf7cc0e994fd15.sql,3,11,1368287
+imdb/benchmarks/ceb/11b/fce5c5de24dbf6034bb0b4698816bc2b19670dce.sql,3,11,1454015
+imdb/benchmarks/ceb/11b/fcb7a4ba993cb02d53b88947e9b9d73964dce26a.sql,3,11,1242451
