+++ workloads/40%-50%/low_variability.csv
@@ -1,5 +1,5 @@
 filepath,num_joins_in_user_query,num_joins_in_benchmark_query,query_id
-imdb/benchmarks/ceb/10a/b0fa6a3f8fa1820a4f7be7aeb6b405909834053c.sql,1,6,38996
-imdb/benchmarks/ceb/8a/8a458.sql,2,11,295874
-imdb/benchmarks/ceb/8a/8a458.sql,2,11,160228
-imdb/benchmarks/ceb/8a/8a458.sql,2,11,13549
+imdb/benchmarks/ceb/10a/ffa89001847857aea332203f58817579edf17efa.sql,1,6,38996
+imdb/benchmarks/ceb/8a/8a99.sql,2,11,295874
+imdb/benchmarks/ceb/8a/8a99.sql,2,11,160228
+imdb/benchmarks/ceb/8a/8a99.sql,2,11,13549
