+++ workloads/40%-50%/mid_variability.csv
@@ -1,42 +1,42 @@
 filepath,num_joins_in_user_query,num_joins_in_benchmark_query,query_id
-imdb/benchmarks/ceb/10a/b0fa6a3f8fa1820a4f7be7aeb6b405909834053c.sql,1,6,118393
-imdb/benchmarks/ceb/10a/b0fa6a3f8fa1820a4f7be7aeb6b405909834053c.sql,1,6,847775
+imdb/benchmarks/ceb/10a/ffa89001847857aea332203f58817579edf17efa.sql,1,6,118393
+imdb/benchmarks/ceb/10a/ffa89001847857aea332203f58817579edf17efa.sql,1,6,847775
 imdb/benchmarks/job/17f.sql,1,6,756887
-imdb/benchmarks/job/17a.sql,1,6,166727
+imdb/benchmarks/job/17e.sql,1,6,166727
 imdb/benchmarks/job/17d.sql,1,6,605945
-imdb/benchmarks/ceb/8a/8a458.sql,3,11,675946
-imdb/benchmarks/job/8a.sql,1,6,83742
-imdb/benchmarks/job/8a.sql,1,6,515257
-imdb/benchmarks/job/8b.sql,1,6,777326
-imdb/benchmarks/job/8b.sql,1,6,654599
-imdb/benchmarks/job/8b.sql,1,6,556000
-imdb/benchmarks/job/8c.sql,1,6,641585
-imdb/benchmarks/job/8c.sql,1,6,325578
-imdb/benchmarks/job/8b.sql,1,6,213057
-imdb/benchmarks/job/17e.sql,1,6,1874094
+imdb/benchmarks/ceb/8a/8a99.sql,3,11,675946
+imdb/benchmarks/job/8d.sql,1,6,83742
+imdb/benchmarks/job/8d.sql,1,6,515257
+imdb/benchmarks/job/8c.sql,1,6,777326
+imdb/benchmarks/job/8c.sql,1,6,654599
+imdb/benchmarks/job/8c.sql,1,6,556000
+imdb/benchmarks/job/8b.sql,1,6,641585
+imdb/benchmarks/job/8b.sql,1,6,325578
+imdb/benchmarks/job/8c.sql,1,6,213057
+imdb/benchmarks/job/17c.sql,1,6,1874094
 imdb/benchmarks/job/17b.sql,1,6,1606052
-imdb/benchmarks/job/17c.sql,1,6,1874684
-imdb/benchmarks/job/8a.sql,1,6,1564691
-imdb/benchmarks/job/8d.sql,1,6,1409300
-imdb/benchmarks/job/8a.sql,1,6,1386532
-imdb/benchmarks/job/8a.sql,1,6,1376963
-imdb/benchmarks/ceb/10a/aa1883d2afa0b06a4d39a9ac7a59f64b70e056bd.sql,1,6,1842498
-imdb/benchmarks/ceb/10a/aa1883d2afa0b06a4d39a9ac7a59f64b70e056bd.sql,1,6,1468494
-imdb/benchmarks/ceb/10a/aa1883d2afa0b06a4d39a9ac7a59f64b70e056bd.sql,1,6,1550638
-imdb/benchmarks/ceb/10a/aa1883d2afa0b06a4d39a9ac7a59f64b70e056bd.sql,1,6,1753688
-imdb/benchmarks/ceb/10a/7bcc6d0a7d4f71732a63dd7ed1fa90b920521629.sql,1,6,1882347
-imdb/benchmarks/ceb/3a/3a710.sql,2,9,1515784
-imdb/benchmarks/ceb/3a/3a710.sql,2,9,1788604
-imdb/benchmarks/ceb/3a/3a1120.sql,2,9,1487916
-imdb/benchmarks/ceb/10a/7bcc6d0a7d4f71732a63dd7ed1fa90b920521629.sql,1,6,1555397
+imdb/benchmarks/job/17a.sql,1,6,1874684
+imdb/benchmarks/job/8d.sql,1,6,1564691
+imdb/benchmarks/job/8a.sql,1,6,1409300
+imdb/benchmarks/job/8d.sql,1,6,1386532
+imdb/benchmarks/job/8d.sql,1,6,1376963
+imdb/benchmarks/job/18c.sql,1,6,1842498
+imdb/benchmarks/job/18c.sql,1,6,1468494
+imdb/benchmarks/job/18c.sql,1,6,1550638
+imdb/benchmarks/job/18c.sql,1,6,1753688
+imdb/benchmarks/job/10c.sql,1,6,1882347
+imdb/benchmarks/ceb/3a/3a999.sql,2,9,1515784
+imdb/benchmarks/ceb/3a/3a999.sql,2,9,1788604
+imdb/benchmarks/ceb/3a/3a998.sql,2,9,1487916
+imdb/benchmarks/job/10c.sql,1,6,1555397
 imdb/benchmarks/job/10b.sql,1,6,1351327
 imdb/benchmarks/job/18b.sql,1,6,1832002
-imdb/benchmarks/job/18c.sql,1,6,1668388
-imdb/benchmarks/ceb/10a/7bcc6d0a7d4f71732a63dd7ed1fa90b920521629.sql,1,6,1407328
-imdb/benchmarks/job/10c.sql,1,6,1440900
+imdb/benchmarks/ceb/10a/ff7e187ff73ce227ca49c7b4791037f242cdbde6.sql,1,6,1668388
+imdb/benchmarks/job/10c.sql,1,6,1407328
+imdb/benchmarks/ceb/10a/ff5392412c047db440be1c25a9368dca0af322a6.sql,1,6,1440900
 imdb/benchmarks/job/10a.sql,1,6,1633354
-imdb/benchmarks/job/10c.sql,1,6,1822184
-imdb/benchmarks/job/10c.sql,1,6,1405833
-imdb/benchmarks/job/18a.sql,1,6,1864422
-imdb/benchmarks/job/18a.sql,1,6,1838741
-imdb/benchmarks/job/10c.sql,1,6,1407790
+imdb/benchmarks/ceb/10a/ff5392412c047db440be1c25a9368dca0af322a6.sql,1,6,1822184
+imdb/benchmarks/ceb/10a/ff5392412c047db440be1c25a9368dca0af322a6.sql,1,6,1405833
+imdb/benchmarks/ceb/10a/ff28e87026770d06e47c8133bde907e8f62af5f2.sql,1,6,1864422
+imdb/benchmarks/ceb/10a/ff28e87026770d06e47c8133bde907e8f62af5f2.sql,1,6,1838741
+imdb/benchmarks/ceb/10a/ff5392412c047db440be1c25a9368dca0af322a6.sql,1,6,1407790
