+++ workloads/50%-60%/low_variability.csv
@@ -1,23 +1,23 @@
 filepath,num_joins_in_user_query,num_joins_in_benchmark_query,query_id
-imdb/benchmarks/ceb/10a/b0fa6a3f8fa1820a4f7be7aeb6b405909834053c.sql,1,6,53026158
-imdb/benchmarks/ceb/10a/b0fa6a3f8fa1820a4f7be7aeb6b405909834053c.sql,1,6,49706397
-imdb/benchmarks/ceb/10a/aa1883d2afa0b06a4d39a9ac7a59f64b70e056bd.sql,1,6,57184297
-imdb/benchmarks/ceb/10a/aa1883d2afa0b06a4d39a9ac7a59f64b70e056bd.sql,1,6,47527818
-imdb/benchmarks/ceb/10a/7bcc6d0a7d4f71732a63dd7ed1fa90b920521629.sql,1,6,50859422
-imdb/benchmarks/ceb/10a/91b99b194361afcf7e32ecc811548342a16e05aa.sql,1,6,42066796
-imdb/benchmarks/ceb/10a/91b99b194361afcf7e32ecc811548342a16e05aa.sql,1,6,42066792
-imdb/benchmarks/ceb/10a/91b99b194361afcf7e32ecc811548342a16e05aa.sql,1,6,42120065
-imdb/benchmarks/ceb/10a/91b99b194361afcf7e32ecc811548342a16e05aa.sql,1,6,56823036
-imdb/benchmarks/ceb/10a/76606f72fdf1a6b058fa85061c5ebaaac1dd0b1f.sql,1,6,56327739
-imdb/benchmarks/ceb/10a/91b99b194361afcf7e32ecc811548342a16e05aa.sql,1,6,53352903
-imdb/benchmarks/ceb/10a/91b99b194361afcf7e32ecc811548342a16e05aa.sql,1,6,42066803
-imdb/benchmarks/ceb/10a/aa1883d2afa0b06a4d39a9ac7a59f64b70e056bd.sql,1,6,43680794
-imdb/benchmarks/ceb/10a/aa1883d2afa0b06a4d39a9ac7a59f64b70e056bd.sql,1,6,47793335
-imdb/benchmarks/ceb/10a/6a8c485255dad8338337788b05933746cfba3a86.sql,1,6,48972429
-imdb/benchmarks/ceb/10a/76606f72fdf1a6b058fa85061c5ebaaac1dd0b1f.sql,1,6,44502026
-imdb/benchmarks/ceb/10a/69392b806b2d245ebf7e4ae196b9947536c45af4.sql,2,11,56027007
-imdb/benchmarks/ceb/10a/d4eec8b3940f97bddc1b30930d723b9b1245aee9.sql,2,11,42120067
-imdb/benchmarks/ceb/10a/af3f46ab118a1c2d503aeeee2e8c5ae79687f8a1.sql,2,11,56944328
-imdb/benchmarks/ceb/10a/69392b806b2d245ebf7e4ae196b9947536c45af4.sql,2,11,48804076
-imdb/benchmarks/ceb/10a/69392b806b2d245ebf7e4ae196b9947536c45af4.sql,2,11,44217733
-imdb/benchmarks/ceb/10a/4f19c56e1f0ccdeca925474a607fb0d9994532ff.sql,2,11,47594268
+imdb/benchmarks/ceb/10a/ffa89001847857aea332203f58817579edf17efa.sql,1,6,53026158
+imdb/benchmarks/ceb/10a/ffa89001847857aea332203f58817579edf17efa.sql,1,6,49706397
+imdb/benchmarks/ceb/10a/ff7e187ff73ce227ca49c7b4791037f242cdbde6.sql,1,6,57184297
+imdb/benchmarks/ceb/10a/ff7e187ff73ce227ca49c7b4791037f242cdbde6.sql,1,6,47527818
+imdb/benchmarks/ceb/10a/ff5392412c047db440be1c25a9368dca0af322a6.sql,1,6,50859422
+imdb/benchmarks/ceb/10a/ff28e87026770d06e47c8133bde907e8f62af5f2.sql,1,6,42066796
+imdb/benchmarks/ceb/10a/ff28e87026770d06e47c8133bde907e8f62af5f2.sql,1,6,42066792
+imdb/benchmarks/ceb/10a/ff28e87026770d06e47c8133bde907e8f62af5f2.sql,1,6,42120065
+imdb/benchmarks/ceb/10a/ff28e87026770d06e47c8133bde907e8f62af5f2.sql,1,6,56823036
+imdb/benchmarks/ceb/10a/feb4a84432cbf3c03659dec9f994e45c20117dce.sql,1,6,56327739
+imdb/benchmarks/ceb/10a/ff28e87026770d06e47c8133bde907e8f62af5f2.sql,1,6,53352903
+imdb/benchmarks/ceb/10a/ff28e87026770d06e47c8133bde907e8f62af5f2.sql,1,6,42066803
+imdb/benchmarks/ceb/10a/ff7e187ff73ce227ca49c7b4791037f242cdbde6.sql,1,6,43680794
+imdb/benchmarks/ceb/10a/ff7e187ff73ce227ca49c7b4791037f242cdbde6.sql,1,6,47793335
+imdb/benchmarks/ceb/10a/feab1842e30c452abd391c3326d9e4ff34c6dc11.sql,1,6,48972429
+imdb/benchmarks/ceb/10a/feb4a84432cbf3c03659dec9f994e45c20117dce.sql,1,6,44502026
+imdb/benchmarks/ceb/10a/fe908e2b6537a8d6acfd277573458dce0e168e28.sql,2,11,56027007
+imdb/benchmarks/ceb/10a/fe466f77ec0cd3dca2c428d45227925db297555a.sql,2,11,42120067
+imdb/benchmarks/ceb/10a/fe2eb3a8d76a12f5a5487e0a328bd294353322ca.sql,2,11,56944328
+imdb/benchmarks/ceb/10a/fe908e2b6537a8d6acfd277573458dce0e168e28.sql,2,11,48804076
+imdb/benchmarks/ceb/10a/fe908e2b6537a8d6acfd277573458dce0e168e28.sql,2,11,44217733
+imdb/benchmarks/ceb/10a/fde055219af175b5eb19a7dc52449920e4abac50.sql,2,11,47594268
