+++ workloads/50%-60%/mid_variability.csv
@@ -1,23 +1,23 @@
 filepath,num_joins_in_user_query,num_joins_in_benchmark_query,query_id
-imdb/benchmarks/ceb/10a/b0fa6a3f8fa1820a4f7be7aeb6b405909834053c.sql,1,6,126264
+imdb/benchmarks/ceb/10a/ffa89001847857aea332203f58817579edf17efa.sql,1,6,126264
 imdb/benchmarks/job/17f.sql,1,6,168179
-imdb/benchmarks/job/8a.sql,1,6,123982
-imdb/benchmarks/job/10b.sql,1,6,156126
-imdb/benchmarks/job/18b.sql,1,6,143303
-imdb/benchmarks/ceb/10a/aa1883d2afa0b06a4d39a9ac7a59f64b70e056bd.sql,1,6,189871
-imdb/benchmarks/ceb/10a/7bcc6d0a7d4f71732a63dd7ed1fa90b920521629.sql,1,6,194130
-imdb/benchmarks/job/10c.sql,1,6,167073
-imdb/benchmarks/ceb/8a/8a458.sql,2,11,183724
-imdb/benchmarks/ceb/8a/8a377.sql,2,11,194676
-imdb/benchmarks/ceb/8a/8a377.sql,2,11,194510
-imdb/benchmarks/ceb/8a/8a377.sql,2,11,179865
-imdb/benchmarks/ceb/8a/8a377.sql,2,11,166205
-imdb/benchmarks/ceb/8a/8a377.sql,2,11,164945
-imdb/benchmarks/ceb/8a/8a377.sql,2,11,194647
-imdb/benchmarks/ceb/8a/8a377.sql,2,11,188694
-imdb/benchmarks/ceb/8a/8a377.sql,2,11,165234
-imdb/benchmarks/ceb/8a/8a377.sql,2,11,186275
-imdb/benchmarks/ceb/8a/8a377.sql,2,11,194010
-imdb/benchmarks/ceb/8a/8a283.sql,2,11,179506
-imdb/benchmarks/ceb/8a/8a377.sql,2,11,191861
-imdb/benchmarks/ceb/8a/8a377.sql,2,11,194804
+imdb/benchmarks/job/8d.sql,1,6,123982
+imdb/benchmarks/job/10c.sql,1,6,156126
+imdb/benchmarks/job/18c.sql,1,6,143303
+imdb/benchmarks/job/8c.sql,1,6,189871
+imdb/benchmarks/job/8b.sql,1,6,194130
+imdb/benchmarks/job/10b.sql,1,6,167073
+imdb/benchmarks/ceb/8a/8a99.sql,2,11,183724
+imdb/benchmarks/ceb/8a/8a98.sql,2,11,194676
+imdb/benchmarks/ceb/8a/8a98.sql,2,11,194510
+imdb/benchmarks/ceb/8a/8a98.sql,2,11,179865
+imdb/benchmarks/ceb/8a/8a98.sql,2,11,166205
+imdb/benchmarks/ceb/8a/8a98.sql,2,11,164945
+imdb/benchmarks/ceb/8a/8a98.sql,2,11,194647
+imdb/benchmarks/ceb/8a/8a98.sql,2,11,188694
+imdb/benchmarks/ceb/8a/8a98.sql,2,11,165234
+imdb/benchmarks/ceb/8a/8a98.sql,2,11,186275
+imdb/benchmarks/ceb/8a/8a98.sql,2,11,194010
+imdb/benchmarks/ceb/8a/8a97.sql,2,11,179506
+imdb/benchmarks/ceb/8a/8a98.sql,2,11,191861
+imdb/benchmarks/ceb/8a/8a98.sql,2,11,194804
