+++ workloads/60%-70%/mid_variability.csv
@@ -2,35 +2,35 @@ filepath,num_joins_in_user_query,num_joins_in_benchmark_query,query_id
 imdb/benchmarks/job/11d.sql,2,7,814577
 imdb/benchmarks/job/11d.sql,2,7,724511
 imdb/benchmarks/job/11d.sql,2,7,750354
-imdb/benchmarks/ceb/10a/b0fa6a3f8fa1820a4f7be7aeb6b405909834053c.sql,1,6,587087
-imdb/benchmarks/ceb/10a/b0fa6a3f8fa1820a4f7be7aeb6b405909834053c.sql,1,6,875262
-imdb/benchmarks/ceb/10a/b0fa6a3f8fa1820a4f7be7aeb6b405909834053c.sql,1,6,785829
+imdb/benchmarks/ceb/10a/ffa89001847857aea332203f58817579edf17efa.sql,1,6,587087
+imdb/benchmarks/ceb/10a/ffa89001847857aea332203f58817579edf17efa.sql,1,6,875262
+imdb/benchmarks/ceb/10a/ffa89001847857aea332203f58817579edf17efa.sql,1,6,785829
 imdb/benchmarks/job/11c.sql,11,11,710782
 imdb/benchmarks/job/11c.sql,11,11,849767
 imdb/benchmarks/job/11c.sql,11,11,761149
-imdb/benchmarks/job/11a.sql,2,7,611253
-imdb/benchmarks/job/11a.sql,2,7,614139
-imdb/benchmarks/job/11a.sql,2,7,850584
-imdb/benchmarks/job/11b.sql,2,7,759332
-imdb/benchmarks/job/11b.sql,2,7,689467
-imdb/benchmarks/job/11b.sql,2,7,789100
-imdb/benchmarks/job/9a.sql,3,7,657038
-imdb/benchmarks/job/9a.sql,3,7,800006
-imdb/benchmarks/job/9a.sql,3,7,767178
+imdb/benchmarks/job/11b.sql,2,7,611253
+imdb/benchmarks/job/11b.sql,2,7,614139
+imdb/benchmarks/job/11b.sql,2,7,850584
+imdb/benchmarks/job/11a.sql,2,7,759332
+imdb/benchmarks/job/11a.sql,2,7,689467
+imdb/benchmarks/job/11a.sql,2,7,789100
+imdb/benchmarks/job/16d.sql,3,7,657038
+imdb/benchmarks/job/16d.sql,3,7,800006
+imdb/benchmarks/job/16d.sql,3,7,767178
 imdb/benchmarks/job/17f.sql,1,6,761675
 imdb/benchmarks/job/17f.sql,1,6,713257
 imdb/benchmarks/job/17f.sql,1,6,853776
-imdb/benchmarks/job/17a.sql,1,6,712676
-imdb/benchmarks/job/17a.sql,1,6,856505
-imdb/benchmarks/job/17a.sql,1,6,858558
-imdb/benchmarks/job/14b.sql,2,7,858559
-imdb/benchmarks/job/14b.sql,2,7,712002
-imdb/benchmarks/job/14b.sql,2,7,802700
-imdb/benchmarks/job/9a.sql,3,7,615193
-imdb/benchmarks/job/9a.sql,3,7,788691
-imdb/benchmarks/job/9a.sql,3,7,757153
-imdb/benchmarks/job/9c.sql,3,7,471867
-imdb/benchmarks/job/9c.sql,3,7,558776
-imdb/benchmarks/job/9c.sql,3,7,569875
-imdb/benchmarks/job/16b.sql,3,7,482256
-imdb/benchmarks/job/16b.sql,3,7,477416
+imdb/benchmarks/job/17e.sql,1,6,712676
+imdb/benchmarks/job/17e.sql,1,6,856505
+imdb/benchmarks/job/17e.sql,1,6,858558
+imdb/benchmarks/job/12c.sql,2,7,858559
+imdb/benchmarks/job/12c.sql,2,7,712002
+imdb/benchmarks/job/12c.sql,2,7,802700
+imdb/benchmarks/job/16d.sql,3,7,615193
+imdb/benchmarks/job/16d.sql,3,7,788691
+imdb/benchmarks/job/16d.sql,3,7,757153
+imdb/benchmarks/job/16c.sql,3,7,471867
+imdb/benchmarks/job/16c.sql,3,7,558776
+imdb/benchmarks/job/16c.sql,3,7,569875
+imdb/benchmarks/job/9d.sql,3,7,482256
+imdb/benchmarks/job/9d.sql,3,7,477416
