+++ workloads/60%-70%/stats.csv
@@ -1,4 +1,4 @@
 workload_type,user_id,instance_id,number_of_queries,query_repetition_rate,n_distinct_num_joins,n_distinct_readsets,n_occurrences_step_2,n_occurrences_step_5,n_occurrences_step_3_to_6,n_occurrences_step_3_to_7,n_occurrences_step_4_to_6,n_occurrences_step_4_to_7
 low_variability,14,54,5,0.600,2,2,0,2,0,0,0,0
 mid_variability,10,50,35,0.686,4,5,5,5,1,0,0,0
-high_variability,17,186,334,0.665,11,35,27,21,18,4,15,27
+high_variability,17,186,334,0.665,11,35,28,21,17,4,15,27
