+++ workloads/80%-90%/low_variability.csv
@@ -1,13 +1,13 @@
 filepath,num_joins_in_user_query,num_joins_in_benchmark_query,query_id
-imdb/benchmarks/ceb/8a/8a458.sql,2,11,1184841
-imdb/benchmarks/ceb/10a/b0fa6a3f8fa1820a4f7be7aeb6b405909834053c.sql,1,6,1414842
-imdb/benchmarks/ceb/8a/8a458.sql,2,11,963093
-imdb/benchmarks/ceb/10a/b0fa6a3f8fa1820a4f7be7aeb6b405909834053c.sql,1,6,571808
-imdb/benchmarks/ceb/8a/8a458.sql,2,11,2624535
-imdb/benchmarks/ceb/8a/8a458.sql,2,11,3724762
-imdb/benchmarks/ceb/10a/b0fa6a3f8fa1820a4f7be7aeb6b405909834053c.sql,1,6,3064434
-imdb/benchmarks/ceb/10a/b0fa6a3f8fa1820a4f7be7aeb6b405909834053c.sql,1,6,3766738
-imdb/benchmarks/ceb/8a/8a458.sql,2,11,2395685
-imdb/benchmarks/ceb/10a/b0fa6a3f8fa1820a4f7be7aeb6b405909834053c.sql,1,6,2597932
-imdb/benchmarks/ceb/8a/8a458.sql,2,11,2789688
-imdb/benchmarks/ceb/10a/b0fa6a3f8fa1820a4f7be7aeb6b405909834053c.sql,1,6,1575030
+imdb/benchmarks/ceb/8a/8a99.sql,2,11,1184841
+imdb/benchmarks/ceb/10a/ffa89001847857aea332203f58817579edf17efa.sql,1,6,1414842
+imdb/benchmarks/ceb/8a/8a99.sql,2,11,963093
+imdb/benchmarks/ceb/10a/ffa89001847857aea332203f58817579edf17efa.sql,1,6,571808
+imdb/benchmarks/ceb/8a/8a99.sql,2,11,2624535
+imdb/benchmarks/ceb/8a/8a99.sql,2,11,3724762
+imdb/benchmarks/ceb/10a/ffa89001847857aea332203f58817579edf17efa.sql,1,6,3064434
+imdb/benchmarks/ceb/10a/ffa89001847857aea332203f58817579edf17efa.sql,1,6,3766738
+imdb/benchmarks/ceb/8a/8a99.sql,2,11,2395685
+imdb/benchmarks/ceb/10a/ffa89001847857aea332203f58817579edf17efa.sql,1,6,2597932
+imdb/benchmarks/ceb/8a/8a99.sql,2,11,2789688
+imdb/benchmarks/ceb/10a/ffa89001847857aea332203f58817579edf17efa.sql,1,6,1575030
